+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fixed legend labels missing when a layer only exists in non-first facets [#721](https://github.com/MakieOrg/AlgebraOfGraphics.jl/pull/721).
+
 ## v0.12.1 - 2026-01-07
 
 - Make the `visual`s `BarPlot`, `BoxPlot`, `Density` and `CrossBar` aware of the `strokecolor` aesthetic. Note that, for some `visual`s, it is necessary to specify a `strokewidth` strictly larger than 0 to appreciate the effect of the `strokecolor` mapping. Note also that the `strokecolor` aesthetic in itself was already available, and it is defined so that it behaves as the `Color` aesthetic (for example, in terms of scales). Indeed, the `visual`s `Scatter`, `ScatterLines` and `Hist` were already aware of `strokecolor`. There are other `visual`s that also could use `strokecolor` (for example, [violin plots](https://docs.makie.org/stable/reference/plots/violin#strokecolor)), but implementing `strokecolor` for all of them is left for later.